pyproject.toml
[tool.poetry]
name = "device-location"
version = "0.1.0"
description = "Device Location project using redis and poetry"
authors = ["HemantMultani <hemant.k.multani@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-redis = "^5.4.0"


[tool.poetry.group.dev.dependencies]
pandas = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
